implement PauliStringTS and speed up OperatorTS2D (#49)
* implement PauliStringTS and speed up OperatorTS2D

Adds PauliStringTS, which represents a translation-symmetric
sum of a Pauli string in N-dimensions.

These sums are constructed by a much more efficient
bit-shift algorithm (credit: ChatGPT), which knows
the periods of the system at compile time.

These improvements are applied to OperatorTS2D.

* remove Val in favor of inlining.

It turns out that Val does not help much and inlining
is enough to propagate at least the Ls to the innermost call.

I also tried precomputing and storing the masks, but it
did not make anything faster. Probably, it is already memory bound.

* fix typo

* refactor OperatorTS1, OperatorTS2 -> OperatorTS

OperatorTS{Ls, U, T} is now a typedef for Operator{PauliStringTS{Ls, U}, T}.
The older constructors are still there but will simply spit out an operator.

Operator(o::OperatorTS;rs) is deprecated in favor of `representative(o)` and `resum(o)`.

`rotate`, `rotate_lower` are deprecated in favor of `shift`
`shift_left`, `shift_origin` and `shift1` are deprecated in favor of simply calling the constructor of `OperatorTS`

* moments: apply trace trick for OperatorTS

* operatorts: make sure find_representative stays inlined…

* operatorts: apply cutoff after binary_kernel

* docs: update to OperatorTS and introduce doctests

* benchmarks: add benchmark for translation symmetry

* docs: lanczos, convert brittle doctests to example blocks